requirements.txt
streamlit
pandas
mysql.connector
matplotlib